requirements.txt
# Core dependencies (compatible with Python 3.9+)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=0.19.0

# CrewAI and LLM (use latest compatible version)
crewai>=0.1.0
langchain-openai>=0.0.5
langchain>=0.1.0

# Data handling
pydantic>=1.10.0

# HTTP requests
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.0

# HTML parsing
beautifulsoup4>=4.11.0
lxml>=4.9.0

# SSE support
sse-starlette>=1.6.0
